:seedling: Create token-agnostic util to get transactions
